Rename import file and class names based on these two formats: - Format: Change (original) obs_instrument_{name} to   (new) obs_volume_{volume/volset}, and for pds4: obs_bundle_{name} - Format: Change (original) obs_mission_{name} to (new)   obs_volume_{name}_common, and for pds4: obs_bundle_{name}_common
