+++ opus/import/obs_volume_cocirs_01xxx.py
@@ -1,7 +1,7 @@
 ################################################################################
-# obs_instrument_cocirs_cube.py
+# obs_volume_cocirs_01xxx.py
 #
-# Defines the ObsInstrumentCOCIRSCube class, which encapsulates fields in the
+# Defines the ObsVolumeCOCIRS01xxx class, which encapsulates fields in the
 # common, obs_mission_cassini, and obs_instrument_cocirs tables for volumes
 # COCIRS_[01]xxx.
 ################################################################################
@@ -10,12 +10,12 @@ import julian
 
 import opus_support
 
-from obs_mission_cassini import ObsMissionCassini
+from obs_volume_cassini_common import ObsVolumeCassiniCommon
 
 _EQUINOX_DATE = julian.tai_from_iso('2009-08-11T01:40:08.914')
 
 
-class ObsInstrumentCOCIRSCube(ObsMissionCassini):
+class ObsVolumeCOCIRS01xxx(ObsVolumeCassiniCommon):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
@@ -314,9 +314,9 @@ class ObsInstrumentCOCIRSCube(ObsMissionCassini):
         return self._index_col('CSS:BODY_SPACECRAFT_RANGE_MIDDLE')
 
 
-    #######################################
-    ### OVERRIDE FROM ObsMissionCassini ###
-    #######################################
+    ############################################
+    ### OVERRIDE FROM ObsVolumeCassiniCommon ###
+    ############################################
 
     # Warning: Both the equi/point/ring index files and the supplemental index files
     # have an OBSERVATION_ID field and they're different! The index file version looks
